+++ lib/src/home/counter/second_page.dart
@@ -1,7 +1,8 @@
 import 'package:lifecycle_example/src/source.dart';
 
 class SecondPage extends StatefulWidget {
-  SecondPage({Key key, this.title = "You're on the Second Page"}) : super(key: key) {
+  SecondPage({Key key, this.title = "You're on the Second Page"})
+      : super(key: key) {
     // ignore: avoid_print
     print(
         '>>>>>>>>>>>>>>>>>>>>>>>>>   SecondPage StatefulWidget being created. Constructor called.');
@@ -15,11 +16,23 @@ class SecondPage extends StatefulWidget {
 class _SecondPageState extends LifecycleState<SecondPage> {
   _SecondPageState() : super() {
     // ignore: avoid_print
-    print(
-        '>>>>>>>>>>>>>>>>>>>>>>>>>   _SecondPageState being created. Constructor called.');
+    print('>>>>>>>>>>>>>>>>>>>>>>>>>   _SecondPageState constructor called.');
   }
-  // Retain the count with a static field.
-  static int _counter = 0;
+
+  CounterBloc bloc;
+
+  @override
+  void initState() {
+    super.initState();
+    bloc = CounterBloc<_SecondPageState>();
+  }
+
+  @override
+  // ignore: file_names
+  void dispose() {
+    bloc.dispose();
+    super.dispose();
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -28,6 +41,14 @@ class _SecondPageState extends LifecycleState<SecondPage> {
     return Scaffold(
       appBar: AppBar(
         title: Text(widget.title),
+        actions: [
+          AppMenu(
+            buttons: [
+              firstPage,
+              thirdPage,
+            ],
+          ),
+        ],
       ),
       body: Center(
         child: Column(
@@ -36,7 +57,7 @@ class _SecondPageState extends LifecycleState<SecondPage> {
             const Text("You're on this second page."),
             const Text('You have pushed the button this many times:'),
             Text(
-              '$_counter',
+              '${bloc.data}',
               style: Theme.of(context).textTheme.headline4,
             ),
           ],
@@ -47,36 +68,47 @@ class _SecondPageState extends LifecycleState<SecondPage> {
           // ignore: avoid_print
           print(
               '>>>>>>>>>>>>>>>>>>>>>>>>>   Second page Button is pressed. setState() called.');
-          setState(() {
-            _counter++;
-          });
+          bloc.onPressed();
         },
         tooltip: 'Increment',
         child: const Icon(Icons.add),
       ),
       persistentFooterButtons: <Widget>[
-        RaisedButton(
-          onPressed: () {
-            Navigator.push(
-              context,
-              MaterialPageRoute<void>(
-                builder: (BuildContext context) => const CountriesList(),
-              ),
-            );
-          },
-          child: const Text('Countries List'),
-        ),
-        RaisedButton(
-          onPressed: () {
-            // ignore: avoid_print
-            print(">>>>>>>>>>>>>>>>>>>>>>>>>   'First Page' button pressed.");
-            Navigator.pop(context);
-          },
-          child: const Text(
-            'First Page',
-          ),
-        ),
+        firstPage,
+        thirdPage,
       ],
     );
   }
+
+  // RaisedButton get countriesList => RaisedButton(
+  //       onPressed: () {
+  //         Navigator.push(
+  //           context,
+  //           MaterialPageRoute<void>(
+  //             builder: (BuildContext context) => CountriesList(),
+  //           ),
+  //         );
+  //       },
+  //       child: const Text('Countries List'),
+  //     );
+
+  RaisedButton get firstPage => backButton(context, label: 'First Page');
+
+  RaisedButton get thirdPage => nextButton(context, ThirdPage(), label: 'Third Page');
+
+      // RaisedButton get thirdPage => RaisedButton(
+      //   onPressed: () {
+      //     // ignore: avoid_print
+      //     print(">>>>>>>>>>>>>>>>>>>>>>>>>   'Third Page' button pressed.");
+      //     Navigator.push(
+      //       context,
+      //       MaterialPageRoute<void>(
+      //         builder: (BuildContext context) => ThirdPage(),
+      //       ),
+      //     );
+      //   },
+      //   child: const Text(
+      //     'Third Page',
+      //   ),
+      // );
 }