+++ lib/src/source.dart
@@ -1,21 +1,23 @@
 
 export 'package:flutter/material.dart';
 
-export 'package:set_state/set_state.dart';
+export 'package:state_set/state_set.dart';
 
 export 'package:lifecycle_example/src/lifecycle/lifecycle_state.dart';
 
 export 'package:lifecycle_example/src/lifecycle/widgets_binding_observer.dart';
 
+export 'package:lifecycle_example/src/lifecycle/raised_button.dart';
+
 export 'package:lifecycle_example/src/infinite_list.dart';
 
 export 'package:lifecycle_example/src/home/lifecycle_example.dart';
 
-export 'package:lifecycle_example/src/home/counter/home_page.dart';
-
 export 'package:lifecycle_example/src/home/counter/first_page.dart';
 
 export 'package:lifecycle_example/src/home/counter/second_page.dart';
+
+export 'package:lifecycle_example/src/home/counter/third_page.dart';
 
 export 'package:lifecycle_example/src/home/countries/countries_list.dart';
 
@@ -27,4 +29,8 @@ export 'package:app_popup_menu/app_popup_menu.dart';
 
 export 'package:lifecycle_example/src/app/menu/app_menu.dart' hide ColorPicker;
 
+export 'package:lifecycle_example/src/home/counter/counter_bloc.dart';
 
+export 'package:lifecycle_example/src/home/counter/counter_bloc.dart';
+
+export 'package:lifecycle_example/src/home/counter/third_bloc.dart';