+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+
+## 0.2.0
+ February 14, 2021
+- Map<String, RaisedButton> menuButtons
+- State_Set from Set_State package
+
 ## 0.1.0 initial commit
  February 07, 2021
 - Setting up the example app.